+++ _samples/advanced_index_actions.go
@@ -7,9 +7,7 @@ package main
 
 import (
 	"context"
-	"crypto/tls"
 	"fmt"
-	"net/http"
 	"os"
 	"strings"
 
@@ -29,12 +27,10 @@ func example() error {
 	client, err := opensearchapi.NewClient(
 		opensearchapi.Config{
 			Client: opensearch.Config{
-				Transport: &http.Transport{
-					TLSClientConfig: &tls.Config{InsecureSkipVerify: true}, // For testing only. Use certificate for validation.
-				},
-				Addresses: []string{"https://localhost:9200"},
-				Username:  "admin", // For testing only. Don't store credentials in code.
-				Password:  "myStrongPassword123!",
+				InsecureSkipVerify: true, // For testing only. Use certificate for validation.
+				Addresses:          []string{"https://localhost:9200"},
+				Username:           "admin", // For testing only. Don't store credentials in code.
+				Password:           "myStrongPassword123!",
 			},
 		},
 	)